Remove unused sheets & styles
