--- styles/hlm.css
@@ -1,360 +1,0 @@
-.chat-message {
-	background: var(--color-bg-dialog);
-}
-.chat-message .message-header {
-	padding-left: 4px;
-	padding-right: 4px;
-	color: var(--color-text-light-heading);
-}
-.hooklineandmecha .window-content {
-	height: 100%;
-	padding: 5px;
-	overflow-y: hidden;
-	background: var(--color-bg-dialog);
-}
-.hooklineandmecha form {
-	height: 100%;
-}
-.hooklineandmecha .flexrow {
-	margin: 5px;
-	display: flex;
-	flex-direction: row;
-}
-.hooklineandmecha .sheet-header .profile-img {
-	flex: 0 0 100px;
-	height: 100px;
-	margin-right: 10px;
-}
-.hooklineandmecha .sheet-header .header-fields {
-	flex: 1;
-	height: 100px;
-}
-.hooklineandmecha .sheet-header h1.charname {
-	height: 40px;
-	padding: 0px;
-	margin: 5px 0;
-	border-bottom: 0;
-}
-.hooklineandmecha .sheet-header h1.charname input {
-	width: 100%;
-	height: 100%;
-	margin: 0;
-	color: var(--color-text-light-heading);
-}
-.hooklineandmecha .sheet-header .type-size-select {
-	display: flex;
-	flex-direction: row;
-	width: 100%;
-	color: var(--color-text-light-heading);
-}
-.hooklineandmecha .sheet-header .type-size-select .dropdown {
-	flex: 0 0 60px;
-	margin-right: 1rem;
-	color: var(--color-text-light-heading);
-}
-.hooklineandmecha .sheet-header .type-size-select .set-types {
-	flex: 0 0 100px;
-	color: var(--color-text-light-heading);
-}
-.hooklineandmecha .resource {
-	width: 100%;
-	height: 40px;
-	float: left;
-	text-align: center;
-	flex: 80%;
-}
-.hooklineandmecha .resource input {
-	width: 28px;
-	height: 28px;
-	color: var(--color-text-light-heading);
-}
-.hooklineandmecha .tabs {
-	flex: 0 0 40px;
-	border-top: 1px solid var(--color-border-light-1);
-	border-bottom: 1px solid var(--color-border-light-1);
-}
-.hooklineandmecha .tabs .item {
-	line-height: 40px;
-	font-weight: bold;
-}
-.hooklineandmecha .tabs .item.active {
-	text-decoration: underline;
-	text-shadow: none;
-}
-.hooklineandmecha .sheet-body {
-	overflow: hidden;
-}
-.hooklineandmecha .sheet-body .tab {
-	height: 100%;
-	overflow-y: auto;
-}
-.hooklineandmecha .editor {
-	min-height: 100px;
-	height: 100%;
-}
-.hooklineandmecha .editor-content {
-	min-height: 100px;
-	height: 100%;
-}
-.hooklineandmecha .attributes {
-	position: relative;
-}
-.hooklineandmecha .attributes .attribute-label {
-	display: flex;
-	justify-content: flex-start;
-	flex-direction: row;
-	flex: 5 0 20px;
-}
-.hooklineandmecha .attributes .attribute-editbox {
-	display: flex;
-	flex: 1 0 20px;
-}
-.hooklineandmecha .attributes .attribute-control {
-	flex: 0 0 20px;
-	text-align: center;
-	line-height: 28px;
-	border: none;
-}
-.hooklineandmecha .attributes .group-control {
-	flex: 0 0 20px;
-	text-align: center;
-	line-height: 28px;
-	border: none;
-}
-.hooklineandmecha .attributes .attribute-roll {
-	flex: 1 0 20px;
-	text-align: center;
-	border-bottom: none;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-.hooklineandmecha .attributes .button {
-	width: 100%;
-	flex: 1;
-	display: block;
-}
-.hooklineandmecha .button {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	margin: 0;
-	background: rgba(0, 0, 0, 0.1);
-	border: 2px groove var(--color-border-light-1);
-	border-radius: 3px;
-	font-size: 14px;
-}
-.hooklineandmecha .button-dark {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	margin: 0;
-	background: var(--color-bg-option);
-	border: 2px groove var(--color-border-light-1);
-	border-radius: 3px;
-	font-size: 14px;
-}
-.hooklineandmecha .button:hover {
-	background-color: #44444466;
-}.hooklineandmecha .button-dark:hover {
-	background-color: #bbbbbb66;
-}
-.hooklineandmecha .fixed-number {
-	text-align: center;
-	color: var(--color-text-dark-inactive);
-}
-.hooklineandmecha .attributes-header {
-	position: sticky;
-	top: 0;
-	left: 0;
-	align-items: center;
-	padding: 0 5px;
-	margin: 5px 0;
-	background: var(--color-bg-header-secondary);
-	border: 1px solid var(--color-border-light-1);
-	border-radius: 2px;
-	text-align: left;
-	font-weight: bold;
-}
-.hooklineandmecha .attributes-list {
-	list-style: none;
-	margin: 0;
-	padding: 0;
-}
-.hooklineandmecha .attributes-list li > * {
-	margin: 0 3px;
-	height: 28px;
-	line-height: 24px;
-	background: transparent;
-	border: none;
-	border-radius: 0;
-	border-bottom: 1px solid var(--color-border-light-1);
-}
-.hooklineandmecha .attributes-flex {
-	display: flex;
-	flex-flow: row wrap;
-	background: transparent;
-	border: none;
-	border-radius: 0;
-	border-bottom: 1px solid var(--color-border-light-1);
-}
-.hooklineandmecha .attributes-flex .flat-wrapbox {
-	margin: 3px;
-	display: flex;
-	border-radius: 0;
-	border: 1px solid var(--color-border-light-1);
-	height: 32px;
-	line-height: 24px;
-	align-items: center;
-	padding-left: 8px;
-}
-.hooklineandmecha .attributes-flex .label-fixed {
-	width: 100px;
-}
-.hooklineandmecha .attributes-flex .editbox-fixed {
-	width: 28px;
-	text-align: center;
-}
-.hooklineandmecha.sheet.actor {
-	min-width: 560px;
-	min-height: 420px;
-}
-.hooklineandmecha.sheet.item {
-	min-width: 460px;
-	min-height: 400px;
-}
-.message-header {
-	background: var(--color-bg-header);
-	font-size: var(--font-size-14);
-	font-weight: bold;
-}
-.flat-display-row {
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	align-items: center;
-	width: 100%;
-	margin-top: 2px;
-	padding-inline: 4px;
-	border-bottom: 1px solid var(--color-border-light-1);
-}
-.flat-display-row .attribute-label {
-	font-size: var(--font-size-12);
-}
-.flat-display-row .attribute-value {
-	font-size: var(--font-size-12);
-	font-weight: bold;
-}
-.attribute-value {
-	text-align: center !important;
-}
-.hit-result {
-	display: flex;
-	flex-direction: row;
-	margin-top: 2px;
-	margin-bottom: 2px;
-	border-top:1px solid var(--color-border-light-1);
-	border-bottom:1px solid var(--color-border-light-1);
-}
-.hit-result .dice-roll {
-	flex: 2 0 0;
-}
-.hit-result .roll-label {
-	display: flex;
-	align-items: center;
-}
-.hit-result .outcome {
-	flex: 2 0 0;
-	justify-content: center;
-}
-.hit-result .descriptor {
-	flex: 3 0 0;
-	justify-content: left;
-}
-.hit-result .roll-outcome-text{
-	font-size: var(--font-size-13);
-	font-weight: bold;
-}
-.import-row {
-	display: flex;
-	flex-direction: row;
-	border: 2px groove var(--color-border-light-1);
-	line-height: 18px;
-}
-.import-row .file-upload {
-	flex-grow: 1;
-	margin: 1px;
-}
-.import-row .file-upload-button {
-	flex: 0 0 70px;
-	height: 20px;
-	font-size: var(--font-size-10);
-}
-.datafile-table {
-	background: rgba(0, 0, 0, 0.1);
-	padding: 0px;
-}
-.datafile-row {
-	display: flex;
-	flex-direction: row;
-	border-bottom: 1px solid var(--color-border-light-1);
-	line-height: 24px;
-}
-.datafile-table .datafile-header {
-	background: var(--color-bg-header);
-	font-size: var(--font-size-14);
-	font-weight: bold;
-	color: var(--color-text-light-heading);
-	justify-content: center;
-}
-.datafile-table .datafile-label {
-	flex: 1 0 200px;
-	margin: 2px;
-	display: flex;
-	padding: 5px;
-	align-items: center;
-}
-.datafile-table .datafile-version {
-	flex: 0 0 50px;
-	margin: 2px;
-	margin-left: 0px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-.datafile-table .datafile-action-button {
-	flex: 0 0 50px;
-	padding: 2px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-.datafile-table .datafile-actions {
-	flex: 0 0 50px;
-	padding: 2px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-#overlay {
-	position: fixed;
-	display: none;
-	width: 100%;
-	height: 100%;
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	background-color: rgba(0,0,0,0.5);
-	z-index: 2;
-	cursor: wait;
-  }
-  .rollable {
-	background: rgba(0, 0, 0, 0.1);
-	border: 2px groove var(--color-border-light-primary);
-	border-radius: 2px;
-  }
-  .rollable:hover {
-	background-color: #44444466;
-    cursor: pointer;
-  }
