+++ styles/hlm.less
@@ -1,183 +1,183 @@
-@background_color_1: #44444466;
+// @background_color_1: #44444466;
 
-.hooklineandmecha {
-	.window-content {
-		height: 100%;
-		padding: 5px;
-		overflow-y: hidden;
-	}
-	form {
-		height: 100%;
-	}
-	.sheet-header {
-		flex: 0 0 100px;
-		overflow: hidden;
-		display: flex;
-		flex-direction: row;
-		flex-wrap: wrap;
-		justify-content: flex-start;
-		margin-bottom: 10px;
-		.profile-img {
-			flex: 0 0 100px;
-			height: 100px;
-			margin-right: 10px;
-		}
-		.header-fields {
-			flex: 1;
-			height: 100px;
-		}
-		h1.charname {
-			height: 40px;
-			padding: 0px;
-			margin: 5px 0;
-			border-bottom: 0;
-			input {
-				width: 100%;
-				height: 100%;
-				margin: 0;
-			}
-		}
-	}
-	.resource {
-		width: 100%;
-		height: 40px;
-		float: left;
-		text-align: center;
-		flex: 80%;
-		input {
-			width: 28px;
-			height: 28px;
-		}
-	}
-	.tabs {
-		flex: 0 0 40px;
-		border-top: 1px solid #aaa;
-		border-bottom: 1px solid #aaa;
-		.item {
-			line-height: 40px;
-			font-weight: bold;
-		}
-		.item.active {
-			text-decoration: underline;
-			text-shadow: none;
-		}
-	}
-	.sheet-body {
-		overflow: hidden;
-		.tab {
-			height: 100%;
-			overflow-y: auto;
-		}
-	}
-	.editor {
-		height: 100%;
-	}
-	.editor-content {
-		height: 100%;
-	}
-	.attributes {
-		position: relative;
-		.attribute-label {
-			display: flex;
-			justify-content: flex-start;
-			flex-direction: row;
-			flex: 5 0 20px;
-		}
-		.attribute-editbox {
-			display: flex;
-			flex: 1 0 20px;
-		}
-		.attribute-control {
-			flex: 0 0 20px;
-			text-align: center;
-			line-height: 28px;
-			border: none;
-		}
-		.group-control {
-			flex: 0 0 20px;
-			text-align: center;
-			line-height: 28px;
-			border: none;
-		}
-		.attribute-roll {
-			flex: 1 0 20px;
-			text-align: center;
-			border-bottom: none;
-			display: flex;
-			align-items: center;
-			justify-content: center;
-		}
-		.button {
-			width: 100%;
-			flex: 1;
-			margin: 0;
-			background: rgba(0, 0, 0, 0.1);
-			border: 2px groove #f0f0e0;
-			border-radius: 3px;
-			font-size: 14px;
-			display: block;
-			&:hover {
-				background-color: @background_color_1;
-			}
-		}
-	}
-	.attributes-header {
-		position: sticky;
-		top: 0;
-		left: 0;
-		align-items: center;
-		padding: 0 5px;
-		margin: 5px 0;
-		background: #cfcdc2;
-		border: 1px solid #aaa;
-		border-radius: 2px;
-		text-align: left;
-		font-weight: bold;
-	}
-	.attributes-list {
-		list-style: none;
-		margin: 0;
-		padding: 0;
-		li {
-			> * {
-				margin: 0 3px;
-				height: 28px;
-				line-height: 24px;
-				background: transparent;
-				border: none;
-				border-radius: 0;
-				border-bottom: 1px solid #aaa;
-			}
-		}
-	}
-	.attributes-flex {
-		display: flex;
-		flex-flow: row wrap;
-		background: transparent;
-		border: none;
-		border-radius: 0;
-		border-bottom: 1px solid #aaa;
-		.flat-wrapbox {
-			margin: 3px;
-			display: flex;
-			border-radius: 0;
-			border: 1px solid #aaa;
-			height: 28px;
-			line-height: 24px;
-		}
-		.label-fixed {
-			width: 100px;
-		}
-		.editbox-fixed {
-			width: 28px;
-		}
-	}
-}
-.hooklineandmecha.sheet.actor {
-	min-width: 560px;
-	min-height: 420px;
-}
-.hooklineandmecha.sheet.item {
-	min-width: 460px;
-	min-height: 400px;
-}
+// .hooklineandmecha {
+// 	.window-content {
+// 		height: 100%;
+// 		padding: 5px;
+// 		overflow-y: hidden;
+// 	}
+// 	form {
+// 		height: 100%;
+// 	}
+// 	.sheet-header {
+// 		flex: 0 0 100px;
+// 		overflow: hidden;
+// 		display: flex;
+// 		flex-direction: row;
+// 		flex-wrap: wrap;
+// 		justify-content: flex-start;
+// 		margin-bottom: 10px;
+// 		.profile-img {
+// 			flex: 0 0 100px;
+// 			height: 100px;
+// 			margin-right: 10px;
+// 		}
+// 		.header-fields {
+// 			flex: 1;
+// 			height: 100px;
+// 		}
+// 		h1.charname {
+// 			height: 40px;
+// 			padding: 0px;
+// 			margin: 5px 0;
+// 			border-bottom: 0;
+// 			input {
+// 				width: 100%;
+// 				height: 100%;
+// 				margin: 0;
+// 			}
+// 		}
+// 	}
+// 	.resource {
+// 		width: 100%;
+// 		height: 40px;
+// 		float: left;
+// 		text-align: center;
+// 		flex: 80%;
+// 		input {
+// 			width: 28px;
+// 			height: 28px;
+// 		}
+// 	}
+// 	.tabs {
+// 		flex: 0 0 40px;
+// 		border-top: 1px solid #aaa;
+// 		border-bottom: 1px solid #aaa;
+// 		.item {
+// 			line-height: 40px;
+// 			font-weight: bold;
+// 		}
+// 		.item.active {
+// 			text-decoration: underline;
+// 			text-shadow: none;
+// 		}
+// 	}
+// 	.sheet-body {
+// 		overflow: hidden;
+// 		.tab {
+// 			height: 100%;
+// 			overflow-y: auto;
+// 		}
+// 	}
+// 	.editor {
+// 		height: 100%;
+// 	}
+// 	.editor-content {
+// 		height: 100%;
+// 	}
+// 	.attributes {
+// 		position: relative;
+// 		.attribute-label {
+// 			display: flex;
+// 			justify-content: flex-start;
+// 			flex-direction: row;
+// 			flex: 5 0 20px;
+// 		}
+// 		.attribute-editbox {
+// 			display: flex;
+// 			flex: 1 0 20px;
+// 		}
+// 		.attribute-control {
+// 			flex: 0 0 20px;
+// 			text-align: center;
+// 			line-height: 28px;
+// 			border: none;
+// 		}
+// 		.group-control {
+// 			flex: 0 0 20px;
+// 			text-align: center;
+// 			line-height: 28px;
+// 			border: none;
+// 		}
+// 		.attribute-roll {
+// 			flex: 1 0 20px;
+// 			text-align: center;
+// 			border-bottom: none;
+// 			display: flex;
+// 			align-items: center;
+// 			justify-content: center;
+// 		}
+// 		.button {
+// 			width: 100%;
+// 			flex: 1;
+// 			margin: 0;
+// 			background: rgba(0, 0, 0, 0.1);
+// 			border: 2px groove #f0f0e0;
+// 			border-radius: 3px;
+// 			font-size: 14px;
+// 			display: block;
+// 			&:hover {
+// 				background-color: @background_color_1;
+// 			}
+// 		}
+// 	}
+// 	.attributes-header {
+// 		position: sticky;
+// 		top: 0;
+// 		left: 0;
+// 		align-items: center;
+// 		padding: 0 5px;
+// 		margin: 5px 0;
+// 		background: #cfcdc2;
+// 		border: 1px solid #aaa;
+// 		border-radius: 2px;
+// 		text-align: left;
+// 		font-weight: bold;
+// 	}
+// 	.attributes-list {
+// 		list-style: none;
+// 		margin: 0;
+// 		padding: 0;
+// 		li {
+// 			> * {
+// 				margin: 0 3px;
+// 				height: 28px;
+// 				line-height: 24px;
+// 				background: transparent;
+// 				border: none;
+// 				border-radius: 0;
+// 				border-bottom: 1px solid #aaa;
+// 			}
+// 		}
+// 	}
+// 	.attributes-flex {
+// 		display: flex;
+// 		flex-flow: row wrap;
+// 		background: transparent;
+// 		border: none;
+// 		border-radius: 0;
+// 		border-bottom: 1px solid #aaa;
+// 		.flat-wrapbox {
+// 			margin: 3px;
+// 			display: flex;
+// 			border-radius: 0;
+// 			border: 1px solid #aaa;
+// 			height: 28px;
+// 			line-height: 24px;
+// 		}
+// 		.label-fixed {
+// 			width: 100px;
+// 		}
+// 		.editbox-fixed {
+// 			width: 28px;
+// 		}
+// 	}
+// }
+// .hooklineandmecha.sheet.actor {
+// 	min-width: 560px;
+// 	min-height: 420px;
+// }
+// .hooklineandmecha.sheet.item {
+// 	min-width: 460px;
+// 	min-height: 400px;
+// }
