+++ styles/sheet.css
@@ -1,4 +1,4 @@
-.hooklineandmecha .sheet-header {
+.sheet-header {
 	flex: 0 0 50px;
 	overflow: hidden;
 	flex-wrap: wrap;
@@ -6,6 +6,18 @@
     padding-left: 5px;
     padding-right: 5px;
 	background: var(--color-bg-header);
+	color: var(--color-text-light-heading);
+}
+.sheet-header h1.charname {
+	height: 40px;
+	padding: 0px;
+	margin: 5px 0;
+	border-bottom: 0;
+}
+.hooklineandmecha .sheet-header h1.charname input {
+	width: 100%;
+	height: 100%;
+	margin: 0;
 	color: var(--color-text-light-heading);
 }
 .actor-sheet-main {
@@ -53,15 +65,8 @@
     flex-grow: 1;
 }
 .attribute-top-row .attribute-button {
-	background: rgba(0, 0, 0, 0.1);
-	border: 2px groove var(--color-border-light-primary);
 	margin: 0;
     padding: 0;
-	border-radius: 2px;
-}
-.attribute-top-row .attribute-button:hover {
-	background-color: #44444466;
-    cursor: pointer;
 }
 .attribute-piece {
     justify-content: space-between;
@@ -299,4 +304,12 @@ input.background {
     padding-bottom: 3px;
     width: 100%;
     border-bottom: 1px solid var(--color-border-highlight);
+}
+.editor {
+	min-height: 100px;
+	height: 100%;
+}
+.editor-content {
+	min-height: 100px;
+	height: 100%;
 }
