+++ styles/utils.css
@@ -1,3 +1,12 @@
+.hooklineandmecha .window-content {
+	height: 100%;
+	padding: 5px;
+	overflow-y: hidden;
+	background: var(--color-bg-dialog);
+}
+.hooklineandmecha form {
+	height: 100%;
+}
 input.sheet-input {
     border: none;
     border-bottom: 1px solid var(--color-border-light-primary);
@@ -11,3 +20,28 @@ input.sheet-input {
     display: flex;
     flex-direction: column;
 }
+#overlay {
+	position: fixed;
+	display: none;
+	width: 100%;
+	height: 100%;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	background-color: rgba(0,0,0,0.5);
+	z-index: 2;
+	cursor: wait;
+  }
+  .btn {
+	background: rgba(0, 0, 0, 0.1);
+	border: 2px groove var(--color-border-light-primary);
+	border-radius: 2px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  .btn:hover {
+	background-color: #44444466;
+    cursor: pointer;
+  }
