+++ templates/fsh-manager.html
@@ -9,10 +9,10 @@
             <li class="datafile-row">
                 <div class="datafile-label">{{file.filename}}</div>
                 <div class="datafile-version">{{file.version}}</div>
-                <div class="button update" filename="{{file.filename}}" version="{{file.version}}" i18n>Update</div>
-                <div class="button remove" filename="{{file.filename}}" version="{{file.version}}" i18n>Delete</div>
+                <div class="btn update" filename="{{file.filename}}" version="{{file.version}}" i18n>Update</div>
+                <div class="btn remove" filename="{{file.filename}}" version="{{file.version}}" i18n>Delete</div>
             </li>
         {{/each}}
     </ol>
-    <div class="button add-new-fsh">Add new...</div>
+    <div class="btn add-new-fsh">Add new...</div>
 </form>
