+++ templates/partials/attribute-box.html
@@ -1,6 +1,6 @@
 <div class="attribute-box flex-row" id="{{key}}">
     <div class="attribute-top-row flex-row">
-        <div class="name-box" attribute="{{key}}" diesize="6" diecount="2">{{label}}</div>
+        <div class="name-box" attribute="{{key}}">{{label}}</div>
         <div class="value-box">{{attribute.total}}</div>
     </div>
     <div class="attribute-bottom-row flex-col">
