--- templates/partials/attributes-tab.html
@@ -1,137 +1,0 @@
-<div class="attributes-header">{{localize "ATTRIBUTETAB.rollable"}}</div>
-<ol class="attributes-list">
-	{{#each attributes.rolled as |attrib key|}}
-	<li class="attributes flexrow">
-		<span class="attribute-label">{{label}}</span>
-		<div class="attribute-editbox flexrow">
-			<input
-				class="attribute-value"
-				type="number"
-				name="system.attributes.rolled.{{key}}.value"
-				value="{{attrib.value}}"
-				data-dtype="Number"
-			/>
-		</div>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="6"
-				diecount="1"
-				attribute="{{key}}"
-			>
-				1d6
-			</div>
-		</div>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="6"
-				diecount="2"
-				attribute="{{key}}"
-			>
-				2d6
-			</div>
-		</div>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="6"
-				diecount="3"
-				attribute="{{key}}"
-			>
-				3d6
-			</div>
-		</div>
-	</li>
-	{{/each}}
-	<li class="attributes flexrow">
-		<span class="attribute-label flexrow"
-			>{{localize "ATTRIBUTETAB.noAttribute"}}</span
-		>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="3"
-				diecount="1"
-				attribute="{{key}}"
-			>
-				1d3
-			</div>
-		</div>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="6"
-				diecount="1"
-				attribute="{{key}}"
-			>
-				1d6
-			</div>
-		</div>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="6"
-				diecount="2"
-				attribute="{{key}}"
-			>
-				2d6
-			</div>
-		</div>
-		<div class="attribute-roll flexrow">
-			<div
-				class="attribute button rollable"
-				diesize="6"
-				diecount="3"
-				attribute="{{key}}"
-			>
-				3d6
-			</div>
-		</div>
-	</li>
-</ol>
-<div class="attributes-header">{{localize "ATTRIBUTETAB.flat"}}</div>
-<div class="attributes-flex">
-	{{#each attributes.flat as |attrib key|}}
-	<div class="flat-wrapbox">
-		<span class="label-fixed">{{label}}</span>
-		<div class="editbox-fixed">
-			<input
-				type="number"
-				name="system.attributes.flat.{{key}}.value"
-				value="{{attrib.value}}"
-				data-dtype="Number"
-			/>
-		</div>
-	</div>
-	{{/each}}
-</div>
-
-<div class="attributes-header">{{localize "ATTRIBUTETAB.ballast"}}</div>
-<div class="attributes-flex">
-	{{#each ballast as |bal key|}}
-		<div class="flat-wrapbox">
-			<span class="label-fixed">{{label}}</span>
-				<div class="editbox-fixed">
-					{{#ifcond key '==' "total"}}
-					<div class="fixed-number">
-						{{bal.value}}
-					</div>
-					{{else}}
-						{{#ifcond key '==' "weight"}}
-						<div class="fixed-number">
-							{{bal.value}}
-						</div>
-						{{else}}
-						<input
-						type="number"
-						name="system.ballast.{{key}}.value"
-						value="{{bal.value}}"
-						data-dtype="Number"
-						/>
-						{{/ifcond}}
-					{{/ifcond}}
-				</div>
-		</div>
-	{{/each}}
-</div>
