+++ templates/roll-dialog.html
@@ -28,5 +28,5 @@
             </li>
         </ol>
     </div>
-    <div class="button">Roll it!</div>
+    <div class="btn">Roll it!</div>
 </form>
