+++ templates/uploader.html
@@ -3,7 +3,7 @@
         <div>Import .fsh</div>
         <div class="import-row">
             <input id="fsh-file-select" type="file" name="fsh-file-select" class="file-upload" accept=".json, .fsh">
-            <div class="file-upload-button button" title="Import" tabindex="-1">
+            <div class="file-upload-button btn" title="Import" tabindex="-1">
                 <label>Import</label>
             </div>
         </div>
